moving SerializationTestUtils form jgrapht-guava to jgrapht-core
* change existing SerializationTest to use SerializationTestUtils
* change existing guava tests to use SerializationTestUtils
